feat(front): User details page
