+++ .github/workflows/cii_front.yml
@@ -20,17 +20,20 @@ jobs:
         uses: actions/setup-node@v4
         with:
           node-version: 22
-          cache: "npm"
-          cache-dependency-path: front/package-lock.json
+          cache: "pnpm"
+          cache-dependency-path: front/pnpm-lock.yaml
+
+      - name: Install pnpm
+        run: corepack enable && corepack prepare pnpm@latest --activate
 
       - name: Install dependencies
-        run: npm ci
+        run: pnpm install --frozen-lockfile
 
       - name: Run Lint
-        run: npm run lint
+        run: pnpm run lint
 
       - name: Run type check
-        run: npm run lint
+        run: pnpm run typecheck
 
       - name: Run Tests
-        run: npm run test
+        run: pnpm run test
