+++ front/package.json
@@ -12,23 +12,23 @@
   },
   "dependencies": {
     "@radix-ui/themes": "^3.2.1",
-    "next": "15.2.4",
-    "radix-ui": "^1.1.3",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
+    "next": "15.3.0",
+    "radix-ui": "^1.2.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "valibot": "^1.0.0",
     "valibot-env": "^0.2.1"
   },
   "devDependencies": {
-    "@eslint/eslintrc": "^3",
-    "@tailwindcss/postcss": "^4",
-    "@types/node": "^20",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
-    "eslint": "^9",
-    "eslint-config-next": "15.2.4",
-    "tailwindcss": "^4",
-    "typescript": "^5",
+    "@eslint/eslintrc": "^3.3.1",
+    "@tailwindcss/postcss": "^4.1.3",
+    "@types/node": "^22.14.1",
+    "@types/react": "^19.1.1",
+    "@types/react-dom": "^19.1.2",
+    "eslint": "^9.24.0",
+    "eslint-config-next": "15.3.0",
+    "tailwindcss": "^4.1.3",
+    "typescript": "^5.8.3",
     "vitest": "^3.1.1"
   },
   "packageManager": "pnpm@9.15.2+sha512.93e57b0126f0df74ce6bff29680394c0ba54ec47246b9cf321f0121d8d9bb03f750a705f24edc3c1180853afd7c2c3b94196d0a3d53d3e069d9e2793ef11f321"
