+++ front/src/types/utils.ts
@@ -1,0 +1,4 @@
+export type PageProps<P extends string> = {
+  params: Promise<{ [slug in P]: string }>;
+  searchParams: Promise<{ [key: string]: string | string[] | undefined }>;
+};
